+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Don't trigger restore_cached_data plugin functionality when assigning the same cached attachment (janko-m)
 
+* Give `Attacher#_promote` and `Attacher#promote` the same method signature (janko-m)
+
 * Add `Attacher#_delete` which now spawns a background job instead of `Attacher#delete!` (janko-m)
 
 * Make `Attacher#cache!`, `Attacher#store!`, and `Attacher#delete!` public (janko-m)